backend/app/image_ai.py
from fastapi import UploadFile
from datetime import datetime, timedelta, timezone
from typing import Dict
import io
from PIL import Image
import logging

logger = logging.getLogger(__name__)

async def analyze_image(image: UploadFile) -> Dict:
    """
    تحليل صورة الشارت واستخراج المعلومات
    """
    try:
        # قراءة الصورة
        contents = await image.read()
        img = Image.open(io.BytesIO(contents))
        
        # التحقق من صيغة الصورة
        if img.format not in ['JPEG', 'PNG', 'GIF', 'BMP']:
            raise ValueError("صيغة الصورة غير مدعومة")
        
        # استخراج المعلومات الأساسية
        # في التطبيق الحقيقي، يتم استخدام نموذج AI لتحليل الصورة
        analysis_data = {
            "symbol": "BTCUSDT",  # يتم استخراجه من الصورة
            "timeframe": "5m",     # يتم استخراجه من الصورة
            "last_candle_time": datetime.now(timezone.utc).replace(tzinfo=None) - timedelta(minutes=5),
            "image_size": img.size,
            "image_format": img.format
        }
        
        logger.info(f"تم تحليل الصورة بنجاح: {analysis_data}")
        return analysis_data
        
    except Exception as e:
        logger.error(f"خطأ في تحليل الصورة: {e}")
        raise

def extract_chart_info(image_path: str) -> Dict:
    """
    استخراج معلومات الشارت من الصورة
    (يمكن توسيع هذه الدالة لاستخدام OCR أو نماذج AI)
    """
    return {
        "symbol": "BTCUSDT",
        "timeframe": "5m",
        "detected_pattern": "Head and Shoulders",
        "confidence": 0.87
    }
