backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
python-dotenv==1.0.0
Pillow==10.1.0
numpy==1.24.3
aiohttp==3.9.1
